иbe-010 x delete project not found (update) (#14)

+++ src/main/java/de/upteams/tasktracker/project/service/impl/ProjectServiceImpl.java
@@ -73,6 +73,10 @@ public class ProjectServiceImpl implements ProjectService {
             throw new RestApiException(HttpStatus.BAD_REQUEST, "Invalid projectId format");
         }
 
+        if (!repository.existsById(projectId)) {
+            throw new ProjectNotFoundException();
+        }
+
         repository.deleteById(projectId);
     }
 }
